Fix `-Werror=dangling-reference` in enum_utils_test.cpp
The problem here is that the call to TestRichEnum1::value_of(19).value() returns a reference to a temporary object, which results in a dangling reference after the end of the expression. To solve this, we need to extend the lifetime of the temporary object so that constexpr and static_assert work correctly.